@@ -11,72 +11,72 @@
     "preview": "vite preview",
     "lint": "CI=true eslint --ignore-pattern '*.d.ts' --ignore-pattern '*.test.*' --ext ts,tsx --max-warnings 0 src",
     "prepare": "husky install",
-    "storybook": "storybook dev -p 6006 -s public",
+    "storybook": "storybook dev -p 6006",
     "test-storybook": "pnpm exec test-storybook --watch",
     "build-storybook": "storybook build"
   },
   "dependencies": {
-    "@chakra-ui/icons": "^2.0.17",
-    "@chakra-ui/react": "^2.5.1",
-    "@emotion/react": "^11.10.5",
-    "@emotion/styled": "^11.10.5",
+    "@chakra-ui/icons": "^2.1.1",
+    "@chakra-ui/react": "^2.8.2",
+    "@emotion/react": "^11.13.0",
+    "@emotion/styled": "^11.13.0",
     "@fontsource/inter": "^4.5.15",
-    "@tanstack/react-query": "^4.24.6",
-    "framer-motion": "^9.0.3",
-    "intl-messageformat": "^10.3.0",
-    "ky": "^0.33.2",
+    "@tanstack/react-query": "^4.36.1",
+    "framer-motion": "^9.1.7",
+    "intl-messageformat": "^10.5.14",
+    "ky": "^0.33.3",
     "lodash-es": "^4.17.21",
-    "query-string": "^8.1.0",
-    "react": "^18.2.0",
-    "react-dom": "^18.2.0",
+    "query-string": "^8.2.0",
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1",
     "react-error-boundary": "^3.1.4",
-    "react-router-dom": "^6.9.0",
-    "zustand": "^4.3.5"
+    "react-router-dom": "^6.25.1",
+    "zustand": "^4.5.4"
   },
   "devDependencies": {
-    "@storybook/addon-a11y": "^7.5.1",
-    "@storybook/addon-actions": "^7.5.1",
-    "@storybook/addon-essentials": "^7.5.1",
-    "@storybook/addon-interactions": "^7.5.1",
-    "@storybook/addon-links": "^7.5.1",
-    "@storybook/addon-mdx-gfm": "^7.5.1",
-    "@storybook/jest": "^0.2.3",
-    "@storybook/react": "^7.5.1",
-    "@storybook/react-vite": "^7.5.1",
-    "@storybook/test-runner": "^0.13.0",
-    "@storybook/testing-library": "^0.2.2",
-    "@testing-library/react-hooks": "^8.0.1",
+    "@chromatic-com/storybook": "^1.6.1",
+    "@storybook/addon-a11y": "^8.2.5",
+    "@storybook/addon-actions": "^8.2.5",
+    "@storybook/addon-essentials": "^8.2.5",
+    "@storybook/addon-interactions": "^8.2.5",
+    "@storybook/addon-links": "^8.2.5",
+    "@storybook/react": "^8.2.5",
+    "@storybook/react-vite": "^8.2.5",
+    "@storybook/test": "^8.2.5",
+    "@storybook/test-runner": "^0.19.1",
+    "@testing-library/react": "^16.0.0",
     "@trivago/prettier-plugin-sort-imports": "4.1.1",
-    "@types/lodash-es": "^4.17.6",
-    "@types/react": "^18.0.24",
-    "@types/react-dom": "^18.0.8",
-    "@types/testing-library__jest-dom": "^5.14.6",
-    "@typescript-eslint/eslint-plugin": "^6.15.0",
-    "@typescript-eslint/parser": "^6.15.0",
-    "@vitejs/plugin-react-swc": "^3.5.0",
+    "@types/lodash-es": "^4.17.12",
+    "@types/react": "^18.3.3",
+    "@types/react-dom": "^18.3.0",
+    "@types/testing-library__jest-dom": "^5.14.9",
+    "@typescript-eslint/eslint-plugin": "^6.21.0",
+    "@typescript-eslint/parser": "^6.21.0",
+    "@vitejs/plugin-react-swc": "^3.7.0",
     "axe-playwright": "^1.2.3",
     "concurrently": "8.2.0",
-    "eslint": "^8.56.0",
-    "eslint-plugin-react": "^7.33.2",
+    "eslint": "^8.57.0",
+    "eslint-plugin-react": "^7.35.0",
     "eslint-plugin-storybook": "^0.6.15",
     "http-server": "^14.1.1",
-    "husky": "^8.0.2",
-    "jest": "^29.3.1",
-    "jsdom": "^21.1.1",
-    "msw": "^1.1.0",
-    "msw-storybook-addon": "^1.9.0",
-    "playwright": "^1.41.1",
+    "husky": "^8.0.3",
+    "jest": "^29.7.0",
+    "jsdom": "^21.1.2",
+    "msw": "^1.3.3",
+    "msw-storybook-addon": "^1.10.0",
+    "playwright": "^1.45.2",
     "prettier": "2.8.6",
-    "pretty-quick": "^3.1.3",
-    "storybook": "^7.5.1",
-    "storybook-addon-react-router-v6": "^2.0.8",
-    "storybook-dark-mode": "^3.0.1",
+    "pretty-quick": "^3.3.1",
+    "storybook": "^8.2.5",
+    "storybook-addon-remix-react-router": "^3.0.0",
+    "storybook-dark-mode": "^4.0.2",
+    "test-runner=0.19.1": "link:@storybook/test-runner=0.19.1",
     "typescript": "^5.3.3",
-    "vite": "^5.0.12",
-    "vite-plugin-checker": "^0.6.2",
-    "vite-tsconfig-paths": "^4.3.1",
-    "vitest": "^1.2.1",
-    "wait-on": "^7.0.1"
+    "vite": "^5.3.4",
+    "vite-plugin-checker": "^0.6.4",
+    "vite-tsconfig-paths": "^4.3.2",
+    "vitest": "^1.6.0",
+    "wait-on": "^7.2.0"
   },
   "browserslist": {
     "production": [
