@@ -1,31 +1,37 @@
-const { mergeConfig } = require("vite");
-const tsconfigPaths = require("vite-tsconfig-paths");
-module.exports = {
-  stories: ["../src/**/*.mdx", "../src/**/*.stories.@(js|jsx|ts|tsx)"],
+import * as tsconfigPaths from "vite-tsconfig-paths";
+
+export default {
+  stories: ["../src/**/*.stories.@(js|jsx|ts|tsx)"],
+
   addons: [
     "@storybook/addon-links",
     "@storybook/addon-essentials",
     "@storybook/addon-interactions",
     "@storybook/addon-a11y",
-    "storybook-addon-react-router-v6",
+    "storybook-addon-remix-react-router",
     "storybook-dark-mode",
-    "@storybook/addon-mdx-gfm",
+    "@chromatic-com/storybook",
   ],
+
   framework: {
     name: "@storybook/react-vite",
     options: {},
   },
+
   features: {
-    storyStoreV7: true,
     interactionsDebugger: true,
   },
+
   async viteFinal(config) {
     return {
       ...config,
       plugins: [...config.plugins, tsconfigPaths.default()],
     };
   },
-  docs: {
-    autodocs: "tag",
+
+  docs: {},
+
+  typescript: {
+    reactDocgen: "react-docgen-typescript",
   },
 };
